Add ENTRYPOINT to plugin builder to build packages
This change adds a script to the plugin_builder Dockerfiles which runs
CMake using a hard-coded iRODS externals path as well as make package.
This could be improved by allowing for variable paths and passing
options through to the cmake and make calls.

In order to use this, run the following, e.g.:

docker run -it --rm \
    -v /path/to/this/repo:/src:ro \
    -v /path/to/build/dir:/bld \
    pam-interactive-builder:ubuntu-20.04

This will result in build artifacts in the host machine in the directory
at /path/to/build/dir. These can then be copied elsewhere to be
installed/tested.

Added a README.md so we don't forget how to build this. :)