[#30] Use plugin operation for native authentication
This plugin checks for and uses the iRODS limited password (stored in scrambled
form in the .irodsA file) in order to avoid reaching out to the PAM service every
time a user wants to authenticate with iRODS. When using this limited password,
the native authentication scheme is used to authenticate the user.

This change adjusts the implementation for using this limited password by using
the "perform_native_auth" operation built into the plugin rather than directly
invoking authenticate_client. This is the way the plugin is using native
authentication after authenticating with PAM. See step_authenticated.