[#16] Fix enabling/disabling SSL
Enabling and disabling SSL before sending sensitive PAM information
over the network either from the client to the server or from the server
to the catalog service provider is essential for security. The plugin
was not doing this correctly. The SSL enabling/disabling logic has been
updated to behave like the pam_password SSL enabling/disabling.