[#33] Replace configuration usage with grid configuration
This commit removes all of the logic around getting values from server_config.json
and checking to make sure that the TTL values are within the configured
boundaries. Instead, the logic used by pam_password for checking configuration
values via R_GRID_CONFIGURATION is used by simply deferring to the database
operation in which the random password is generated.