[#40] Add insecure_mode configuration
This commit adds a configuration to server_config for pam_interactive
which allows the user to authenticate without SSL enabled. This config
can be set like this in server_config.json:

{
    "plugin_configuration": {
        "authentication": {
            "pam_interactive": {
                "insecure_mode": false
            }
        }
    }
}

The value defaults to false and must be set to true in order to enable
insecure mode. Once insecure mode is active, warning messages will appear
in the server logs as a reminder to the administrator that sensitive user
information is being sent over the network without encryption. The warning
messages only appear when SSL is not enabled and insecure_mode is enabled.