[#37] Add PAM stack configuration
This commit adds a new configuration for pam_interactive which instructs
the server-side plugin on which PAM stack file to use. This PAM stack
file defaults to "irods" (as this has been the historical choice, hard-
coded in the plugin). The configuration will look for the file in /etc/pam.d
as per usual. Here is how to configure the PAM stack name in server_config:

{
    "plugin_configuration": {
        "authentication": {
            "pam_interactive": {
                "pam_stack_name": "irods"
            }
        }
    }
}

This commit also refactors the server configuration fetching functions to
be a little more streamlined.