[#47] Fix authenticating with catalog service consumer
When a client is connected to a catalog service consumer and wishes
to authenticate using pam_interactive, the server is forced to redirect
to a catalog service provider in order to complete the operations because
the end result of authenticating with this plugin is generating a password
to use for authenticating with iRODS. pam_interactive must maintain state
with a PAM session in order to complete all of the steps of a particular
authentication flow as configured by the system administrator. In order to
do this, the session must be managed in a single iRODS agent. As such, the
catalog service provider must be the agent managing the state as the session
is the only way that the server knows the user is authenticatd with PAM and
therefore is allowed to authenticate with iRODS.

There was an issue when redirecting connections to the catalog service
provider wherein clients would get stuck in an infinite loop - specifically,
on the first step of the authentication flow - because the plugin was
disconnecting from the catalog service consumer every time it finished a
step. This would cause the agent to be torn down and the state of the PAM
session was lost, causing it to start all over again when the client was
ready to go to the next step.

This change removes the call to disconnect from the catalog service provider
on redirect so that the connection, and, therefore, the agent, will be reused
for the next step and the PAM session state will be retained.