[#47] Add redirect to pam_auth_agent_request
Redirection to the catalog service provider is required by this plugin
in order to function. This redirection is happening in pam_auth_agent_response
because that is the operation where the plugin advances and interprets
messages from the PAM session conversation.

This change also adds redirection to pam_auth_agent_request so that the
plugin is configured in the agent where the PAM authentication is happening.
This sets things such as the RsComm authentication scheme and the log level
for the pam_interactive plugin's log category.

This change also adds some additional explanatory code comments to give
more information about the redirection logic.