[#38] Fix client-side password expiration detection
The client-side plugin has a feature for user convenience which allows
the plugin to display the configured PAM authentication flow when the
user's authentication has expired, even when running regular iCommands.
Normally, an iRODS error will be displayed because the randomly generated
password is checked with the server.

The fix involves using the configured password_min_time in R_GRID_CONFIGURATION
instead of 0 when no TTL is passed in. This would be the TTL being used
by the server, so the guesswork is being taken out of the process. Due to
this behavior being in the plugin, some logic from the iRODS database plugin
has been duplicated in order to return certain values in the face of
configuration errors in order to match the behavior.