[#38] Remove state file feature
The state file was used to store information about the PAM session so
that certain fields could be given default values when re-authenticating
as well as tracking the expiration of a PAM session to give a better
user experience. These features have since been removed, and now nothing
is being written to this file. This commit removes the logic which
creates and reads from the file.